Add files via upload
tools and technologies-
IDE-eclipse
DATA BASE-MySQL
SERVER-apache tomcat 10

JSP(JavaServer Pages): It is a technology that can enable the creation of dynamic web pages using Java and it can allow mixing the static HTML content with dynamic Java code. This file has .jsp extension and is translated into the servlets by the web container before being executed.
Registration Form: The web form used to collect user information such as the name, phone number, address, etc. Typically for the registration process and these commonly on the websites and applications that are required the user accounts.
Scriptlet: It is a block of the java code that can be embedded with a JSP page with the delimited by the <% %> tags.
Request Parameter: The request parameter sent from the client to the server as part of the HTTP request. In the registration form, form inputs fields values are sent as the request parameters to the server when the form will be submitted.
Server-side Processing: The processing of the data that can occur on the server after receiving the request from the client. The server-side processing involves tasks such as data validation database interaction and generating the response to send back to the client.